module7-genai-langchain/requirements.txt
langchain 
langchain_community 
langchain-openai 
scikit-learn 
langchain-ollama  
requests
ipykernel
chromadb
bs4
python-dotenv
ipywidgets
langgraph-checkpoint-sqlite
langgraph-checkpoint-postgres
tavily
sounddevice
pypdf
fastmcp

